Rename '/wards/schedule-success' to '/wards/book-a-visit-success'
