+++ pageTests/wards/book-a-visit-success.test.js
@@ -1,11 +1,11 @@
-import { getServerSideProps } from "../../pages/wards/schedule-success";
+import { getServerSideProps } from "../../pages/wards/book-a-visit-success";
 //
 // TODO: This needs to be moved once the verifyToken logic is in the container..
 jest.mock("../../src/usecases/userIsAuthenticated", () => () => (token) =>
   token && { ward: "123" }
 );
 
-describe("wards/schedule-success", () => {
+describe("wards/book-a-visit-success", () => {
   const anonymousReq = {
     headers: {
       cookie: "",
